Rename `token` pkg to `auth`
