+++ .golangci.yml
@@ -1,11 +1,20 @@
 linters:
   fast: false
   enable:
+    - misspell
     - gosec
+    - godot
+    - revive
     - errorlint
 linters-settings:
-  staticcheck:
-    go: 1.17
-issues:
-  exclude:
-    - SA3000
+  errcheck:
+    check-type-assertions: true
+  misspell:
+    locale: US
+    ignore-words: []
+  revive:
+    rules:
+      - name: unexported-return
+        disabled: true
+      - name: exported
+        disabled: false
